# storage/api-docs/api-docs.json
{"openapi":"3.0.0","info":{"title":"Beachcomber Api Documentation","description":"How to use Beachcomber Api","contact":{"name":"Beachcomber","email":"mauricia@beachcomber.com"},"license":{"name":"Terms and Condition","url":"https:\/\/www.beachcomber-hotels.com\/fr\/terms-conditions"},"version":"1.0.0"},"servers":[{"url":"http:\/\/localhost:8000","description":"Local API Server"}],"paths":{"\/api\/v2\/auth\/login":{"post":{"tags":["Authentification"],"summary":"User Login","description":"Login User Here","operationId":"authLogin","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"requestBody":{"content":{"multipart\/form-data":{"schema":{"required":["user_email","user_password"],"properties":{"user_email":{"type":"email","example":"example@beachcomber.com"},"user_password":{"type":"password","example":"beachcomber"}},"type":"object"}},"application\/json":{"schema":{"required":["user_email","user_password"],"properties":{"user_email":{"type":"email","example":"example@beachcomber.com"},"user_password":{"type":"password","example":"beachcomber"}},"type":"object"}}}},"responses":{"200":{"description":"Login Successfully","content":{"application\/json":{"schema":{"properties":{"access_token":{"type":"string","example":"string"},"token_type":{"type":"string","example":"string"},"expires_in":{"type":"integer","example":"360"},"user":{"type":"object","example":"{}"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}}}},"\/api\/v2\/auth\/register":{"post":{"tags":["Authentification"],"summary":"User Register","description":"User Register here","operationId":"Register","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"requestBody":{"content":{"multipart\/form-data":{"schema":{"required":["user_email","user_password","user_password_confirmation","user_surname"],"properties":{"user_surname":{"type":"text","example":"beachcomber"},"user_email":{"type":"email","example":"example@beachcomber.com"},"user_password":{"type":"password","example":"beachcomber"},"user_password_confirmation":{"type":"password","example":"beachcomber"}},"type":"object"}},"application\/json":{"schema":{"required":["user_email","user_password","user_password_confirmation","user_surname"],"properties":{"user_surname":{"type":"text","example":"beachcomber"},"user_email":{"type":"email","example":"example@beachcomber.com"},"user_password":{"type":"password","example":"beachcomber"},"user_password_confirmation":{"type":"password","example":"beachcomber"}},"type":"object"}}}},"responses":{"201":{"description":"User created successfully","content":{"application\/json":{"schema":{"properties":{"successMsg":{"type":"string","example":"string"},"data":{"$ref":"#\/components\/schemas\/User"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}}}},"\/api\/v2\/auth\/me":{"get":{"tags":["Authentification"],"summary":"Get the logged in user","description":"Returns current user","operationId":"getProfile","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"responses":{"200":{"description":"User successfully collects","content":{"application\/json":{"schema":{"properties":{"successMsg":{"type":"string","example":"string"},"data":{"$ref":"#\/components\/schemas\/User"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}},"security":[{"bearer":[]}]}},"\/api\/v2\/auth\/update":{"put":{"tags":["Authentification"],"summary":"Update of the authenticated user","description":"User update here","operationId":"authenticatedUpdate","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"requestBody":{"content":{"multipart\/form-data":{"schema":{"required":["user_surname"],"properties":{"user_surname":{"type":"string","example":"string"},"user_othername":{"type":"string","example":"string"},"user_jobtitle":{"type":"string","example":"string"},"user_phone":{"type":"string","example":"string"},"user_name":{"type":"string","example":"string"}},"type":"object"}},"application\/json":{"schema":{"required":["user_surname"],"properties":{"user_surname":{"type":"string","example":"string"},"user_othername":{"type":"string","example":"string"},"user_jobtitle":{"type":"string","example":"string"},"user_phone":{"type":"string","example":"string"},"user_name":{"type":"string","example":"string"}},"type":"object"}}}},"responses":{"200":{"description":"User updated successfully","content":{"application\/json":{"schema":{"properties":{"successMsg":{"type":"string","example":"string"},"data":{"$ref":"#\/components\/schemas\/User"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}},"security":[{"bearer":[]}]}},"\/api\/v2\/auth\/logout":{"post":{"tags":["Authentification"],"summary":"User Logout","description":"Logout User Here","operationId":"authLogout","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"responses":{"200":{"description":"Logout Successfully","content":{"application\/json":{"schema":{"properties":{"successMsg":{"type":"string","example":"string"},"data":{"type":"object","example":"null"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}},"security":[{"bearer":[]}]}},"\/api\/v2\/auth\/refresh":{"post":{"tags":["Authentification"],"summary":"Refresh token","description":"Refresh Token Here","operationId":"authRefresh","parameters":[{"name":"lang","in":"path","required":true,"schema":{"type":"string"},"example":"en"}],"responses":{"200":{"description":"Refresh Successfully","content":{"application\/json":{"schema":{"properties":{"access_token":{"type":"string","example":"string"},"token_type":{"type":"string","example":"string"},"expires_in":{"type":"integer","example":"360"},"user":{"type":"object","example":"{}"}},"type":"object"}}}},"400":{"description":"Bad request","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"401":{"description":"Unauthorized","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}},"404":{"description":"Resource Not Found","content":{"application\/json":{"schema":{"properties":{"errNo":{"type":"integer","example":"number"},"errMsg":{"type":"string","example":"string"}},"type":"object"}}}}},"security":[{"bearer":[]}]}}},"components":{"schemas":{"User":{"required":["user_email","user_password","user_password_confirmation","user_surname"],"properties":{"user_id":{"type":"integer","example":"number"},"user_type_id":{"type":"integer","example":"number"},"user_group_id":{"type":"integer","example":"number"},"user_surname":{"type":"string","example":"string"},"user_othername":{"type":"string","example":"string"},"user_email":{"type":"string","example":"string"},"user_jobtitle":{"type":"string","example":"string"},"user_phone":{"type":"string","example":"string"},"user_name":{"type":"string","example":"string"},"user_password":{"type":"string","example":"string"},"active":{"type":"integer","example":"number"}},"type":"object","xml":{"name":"User"}}},"securitySchemes":{"bearer":{"type":"http","description":"Authorization token obtained from logging in.","name":"Authorization","in":"header","scheme":"bearer"}}},"tags":[{"name":"Authentification","description":"API Endpoints of Authentification"},{"name":"Users","description":"API Endpoints of User"},{"name":"Hotels","description":"API Endpoints of Hotels"},{"name":"Rooms","description":"API Endpoints of Rooms"},{"name":"Tariffs","description":"API Endpoints of Tariffs"}]}